requirements.txt
Flask
python-dotenv
openai
pypdf2
sarvamai
requests